requirements.txt
flask
torch
torchvision
timm
opencv-python-headless
pillow
numpy
